+++ app/src/main/java/com/example/ecomarket/ui/screens/CartScreen.kt
@@ -129,7 +129,7 @@ fun CartItemRow(
 // ---------- Bottom Bar ----------
 @Composable
 fun CartBottomBar(
-    total: Double,
+    total: Int,
     onContinueClick: () -> Unit,
     isEnabled: Boolean
 ) {
