+++ app/src/main/java/com/example/ecomarket/ui/viewmodel/ProductsViewModel.kt
@@ -57,7 +57,7 @@ class ProductsViewModel(
         }
     }
 
-    fun getCartTotal(): Double {
+    fun getCartTotal(): Int {
         return _uiState.value.cartItems.sumOf { it.product.price * it.quantity }
     }
